+++ src/main/java/com/dic/bill/dao/ObjParDAO.java
@@ -9,6 +9,6 @@ import org.springframework.data.repository.query.Param;
 public interface ObjParDAO extends JpaRepository<ObjPar, Integer> {
 
 	@Query("select t from ObjPar t where t.ko.id = :klskId and t.lst.cd = :cd")
-	ObjPar getByKlskCd(@Param("klskId") Integer klskId, @Param("cd") String cd);
+	ObjPar getByKlskCd(@Param("klskId") Long klskId, @Param("cd") String cd);
 
 }
